requirements.txt
streamlit==1.31.0
python-dotenv==1.0.0
openai==1.31.0
tiktoken
langchain==0.2.3
langchain-core==0.2.11
langchain-community==0.2.4
langchain-openai==0.1.8
unstructured==0.9.2
faiss-cpu==1.12.0
python-magic==0.4.27
typing_extensions>=4.7.0
pydantic>=2.4.0
pillow>=10.3.0
nltk
